% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfunctions.R
\name{calc_scags}
\alias{calc_scags}
\title{Compute selected scagnostics}
\usage{
calc_scags(
  x,
  y,
  scags = c("outlying", "stringy", "striated", "clumpy", "sparse", "skewed", "convex",
    "skinny", "monotonic", "splines", "dcor")
)
}
\description{
Compute selected scagnostics
}
\examples{

}
