% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_clumpy}
\alias{sc_clumpy}
\title{Compute clumpy scagnostic measure using MST}
\usage{
sc_clumpy(x, y)
}
\description{
Compute clumpy scagnostic measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_clumpy(anscombe$x1, anscombe$y1)
  sc_clumpy(anscombe$x2, anscombe$y2)
  sc_clumpy(anscombe$x3, anscombe$y3)
  sc_clumpy(anscombe$x4, anscombe$y4)

}
