% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{sc_dcor}
\alias{sc_dcor}
\title{Distance correlation index.}
\usage{
sc_dcor(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
(Taken from tourr git repo)
Computes the distance correlation based index on
2D projections of the data.
}
\keyword{hplot}
