% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{sc_monotonic}
\alias{sc_monotonic}
\title{Measure of Spearman Correlation}
\usage{
sc_monotonic(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
double
}
\description{
Measure of Spearman Correlation
}
