% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfunctions.R
\name{sc_pairwise}
\alias{sc_pairwise}
\title{Compute scagnostics on all possible scatter plots for the given data}
\usage{
sc_pairwise(
  all_data,
  scags = c("outlying", "stringy", "striated", "clumpy", "sparse", "skewed", "convex",
    "skinny", "monotonic", "splines", "dcor"),
  groups = NULL
)
}
\description{
Compute scagnostics on all possible scatter plots for the given data
}
\examples{
#calculate selected scagnsotics
sc_pairwise(datasaurus_dozen_wide, scags=c("outlying","clumpy","monotonic"))

#calculate all scagnsotics for all variable pairs
sc_pairwise(datasaurus_dozen_wide)

}
