% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{sc_splines}
\alias{sc_splines}
\title{Spline based index.}
\usage{
sc_splines(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
(Taken from tourr git repo)
Compares the variance in residuals of a fitted
spline model to the overall variance to find
functional dependence in 2D projections
of the data.
}
\examples{

}
