% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfunctions.R
\name{sc_splom}
\alias{sc_splom}
\title{Build a SPLOM for the selected scagnostics}
\usage{
sc_splom(sca_dataset, interactive = FALSE)
}
\description{
Build a SPLOM for the selected scagnostics
}
\examples{
scag_data <- sc_pairwise(datasaurus_dozen_wide)
sc_splom(scag_data)
}
