% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree.R
\name{scree}
\alias{scree}
\title{Generate an scatter tree}
\usage{
scree(x, y, binner = NULL, ...)
}
\arguments{
\item{x, y}{numeric vectors}

\item{...}{other args}

\item{num_bins}{(default = 30)}
}
\value{
A list of graph objects with three components
 - minimum spanning tree
 - convex hull
 - concave hull (alpha hull)
}
\description{
Generate an scatter tree
}
